backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
numpy
scikit-learn
tqdm
openai
pyyaml